backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
boto3==1.34.0
Pillow==10.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
dj-database-url==2.1.0
django-storages==1.14.2